Code review on ListView parts
OK, no more commits to master. This is the last one in this fashion.
The only freedom I took in the review was to keep the Ids of
layout elements as the first attribute. I also removed the GridView
which was not used in any way. Other then that just some minor cleanup.

We should probably also write a few more words in our commit messages,
at least the ones that are being prepared for code review. Maybe we
should even make a habit of squashing the commits in a feature branch by
rebasing or something fancy.